Moved performance test files to separate directory in tests, renamed performance.py to run_perf.py, updated docs for this change, split the actual test configurations into separate perf_tests.py file for way to access tests info without running from a spheral install
