+++ cmake/SetupSpheral.cmake
@@ -175,9 +175,6 @@ if (SPHERAL_ENABLE_TESTS)
 
   include(${SPHERAL_ROOT_DIR}/cmake/spheral/SpheralInstallPythonFiles.cmake)
   spheral_install_python_tests(${SPHERAL_ROOT_DIR}/tests/ ${SPHERAL_TEST_INSTALL_PREFIX})
-  # Always install performance.py in the top of the testing script
-  install(FILES ${SPHERAL_ROOT_DIR}/tests/performance.py
-    DESTINATION ${SPHERAL_TEST_INSTALL_PREFIX})
 endif()
 
 include(${SPHERAL_ROOT_DIR}/cmake/SpheralConfig.cmake)
