+++ docs/developer/dev/diagnostic_tools.rst
@@ -140,14 +140,14 @@ Performance Regression Testing
 .. note::
    The following is currently only applicable for use on LC machines at LLNL.
 
-``tests/performance.py`` contains a set of performance regression tests. These tests allow a developer to estimate the performance implications of code under development and compare it to the current development branch of Spheral.
+``tests/run_perf.py`` contains a set of performance regression tests. These tests allow a developer to estimate the performance implications of code under development and compare it to the current development branch of Spheral.
 When a merge to the develop branch occurs, the CI runs this regression test multiple times to accumulate benchmark timing data in a centralized directory (``/usr/WS2/sduser/Spheral/benchmark``).
 The general procedure to comparing performance regression tests is:
 
 * Run the performance regression tests from an installation using 2 nodes (number of nodes used in benchmark run by CI):
    ::
 
-      ./spheral-ats --log test_dir_name --numNodes 2 tests/performance.py
+      ./spheral-ats --log test_dir_name --numNodes 2 tests/run_perf.py
 
    There is also a ``--threads`` option to specify a given number of threads per rank.
    The test above will create an ATS python file, ``test_dir_name/atsr.py``, as well as
@@ -162,7 +162,7 @@ The general procedure to comparing performance regression tests is:
 
       ./spheral ./scripts/performance_analysis.py --perfdata test_dir_name --ref /directory/of/reference/caliper/files/
 
-   The input to ``--ref`` can be also be an ATS directory created from running ``performance.py`` or just a directory of Caliper files.
+   The input to ``--ref`` can be also be an ATS directory created from running ``run_perf.py`` or just a directory of Caliper files.
    Removing the ``--ref`` input will default to comparing to benchmark timings in ``/usr/WS2/sduser/Spheral/benchmark``.
    Timing trees can be displayed using the ``--display`` flag.
    The script above computes the mean (:math:`\mu`) and standard deviation (:math:`\sigma`) of the inclusive average time per rank (``Avg time/rank``) timers for each test in the reference (or benchmark) data.
@@ -180,7 +180,7 @@ The general procedure to comparing performance regression tests is:
 
 .. note::
 
-   If ``performance.py`` is run on a non-MPI Spheral build, it will only use 1 rank and will thread all other available cores.
+   If ``run_perf.py`` is run on a non-MPI Spheral build, it will only use 1 rank and will thread all other available cores.
 
 Historical Timing Benchmarks
 ----------------------------
