+++ tests/CMakeLists.txt
@@ -3,3 +3,4 @@ add_subdirectory(cpp/Geometry)
 add_subdirectory(cpp/Utilities)
 add_subdirectory(cpp/Kernel)
 add_subdirectory(cpp/Neighbor)
+add_subdirectory(performance)
