+++ tests/performance/CMakeLists.txt
@@ -1,0 +1,4 @@
+
+# Always install performance.py in the top of the testing script
+install(FILES run_perf.py perf_tests.py
+  DESTINATION ${SPHERAL_TEST_INSTALL_PREFIX})
