+++ tests/performance/run_perf.py
@@ -1,0 +1,150 @@
+#!/user/bin/env python3
+
+# This file runs and compares performance tests through the ats system.
+# Run using: ./spheral-ats tests/run_perf.py
+
+import sys, shutil, os, time, stat
+import numpy as np
+import SpheralConfigs
+import perf_tests as pt
+from Spheral import TimerMgr
+from SpheralTestUtilities import num_3d_cyl_nodes
+from ats import configuration
+
+if (not TimerMgr.timers_usable()):
+    log("WARNING: Timers not enabled, skipping performance tests. Configure Spheral w/ -DSPHERAL_ENABLE_TIMERS=On", echo=True)
+    sys.exit(0)
+
+# Get options from ats
+opts = getOptions()
+
+# Adding --threads to the command line arguments of spheral-ats
+# can force performance to use multiple threads per rank
+num_threads = 1
+if "threads" in opts:
+    num_threads = opts["threads"]
+
+# Adding --ciRun to the command line arguments of spheral-ats
+# triggers copy of Caliper files to benchmark location
+benchmark_dir = None
+test_runs = 1 # Number of times to run each test
+CIRun = False
+if "cirun" in opts and opts["cirun"]:
+    CIRun = True
+    test_runs = 5
+    benchmark_dir = opts["benchmark_dir"]
+is_rerun = False
+if "rerun" in opts and opts["rerun"]:
+    is_rerun = True
+#---------------------------------------------------------------------------
+# Hardware configuration
+#---------------------------------------------------------------------------
+# This should be {$SYS_TYPE}_{compiler name}_{compiler version}_{mpi or cuda info}
+spheral_install_config = SpheralConfigs.config()
+mpi_enabled = SpheralConfigs.enable_mpi()
+# Retrieve the host name and remove any numbers
+temp_uname = os.uname()
+hostname = temp_uname[1].rstrip("0123456789")
+mac_procs = {"rzhound": 112, "rzwhippet": 112, "dane": 112,
+             "rzadams": 84, "rzvernal": 64, "tioga": 64,
+             "rzansel": 40, "lassen": 40, "rzgenie": 36}
+# Find out how many nodes our allocation has grabbed
+num_nodes = max(1, configuration.machine.numNodes)
+if (not mpi_enabled):
+    if (num_nodes > 1):
+        raise Exception("Should not use more than 1 node when MPI is off")
+
+total_num_cores = 0
+try:
+    total_num_cores = mac_procs[hostname] * num_nodes
+except:
+    log("Machine name not recognized", echo=True)
+    raise Exception
+# If MPI is turned off, thread the whole node
+if (not mpi_enabled):
+    num_cores = int(total_num_cores)
+else:
+    # Ideally, tests should be run with 2 nodes and each test will
+    # use one entire node, except the 2D tests which use half a node
+    num_cores = int(total_num_cores/2)
+
+def gather_files(manager):
+    '''
+    Function to gather Caliper file when ATS is finished running.
+    Used by ATS for gathering benchmark Caliper files.
+    '''
+    instpath = os.path.join(benchmark_dir, spheral_install_config)
+    macpath = os.path.join(instpath, hostname)
+    outdir = os.path.join(macpath, "latest")
+    if (not is_rerun):
+        if (os.path.exists(outdir)):
+            # Move existing benchmark data to a different directory
+            log(f"Renaming existing {outdir} to {int(time.time())}", echo=True)
+            os.rename(outdir, os.path.join(macpath, f"{int(time.time())}"))
+        log(f"Creating {outdir}", echo=True)
+        os.makedirs(outdir)
+    filtered = [test for test in manager.testlist if test.status is PASSED]
+    # Set read/write/execute permissions for owner and group
+    perms = stat.S_IRWXU | stat.S_IRWXG
+    for test in filtered:
+        run_dir = test.directory
+        cali_filename = test.options["caliper_filename"]
+        cfile = os.path.join(run_dir, cali_filename)
+        test_name = test.options["label"]
+        outfile = os.path.join(outdir, cali_filename)
+        log(f"Copying {cali_filename} to {outdir}", echo=True)
+        if (CIRun):
+            shutil.copy(cfile, outfile)
+            os.chmod(outfile, perms)
+            shutil.chown(outfile, group="sduser")
+    if (CIRun):
+        cpaths = [outdir, macpath, instpath, benchmark_dir]
+        for p in cpaths:
+            os.chmod(p, perms)
+            shutil.chown(p, group="sduser")
+
+def spheral_setup_test(test_param, ncores, threads=1, **kwargs):
+    '''
+    General method for creating an individual performance test
+    Parameters:
+    test_param: TestParam class for specific test
+    ncores: Total number of cores to use for the test, not number of ranks
+    threads: Number of threads per rank
+    **kwargs: Any additional keyword arguments to pass to ATS tests routine
+    '''
+    if (not mpi_enabled):
+        threads = ncores
+        ncores = 1
+    test_file = test_param.test_file()
+    tests = test_param.get_tests(NCores)
+    for test_name, inps in tests.items():
+        for i in range(test_runs):
+            if (test_runs > 1):
+                cali_name = f"{test_name}_{i}_{int(time.time())}.cali"
+            else:
+                cali_name = f"{test_name}_{int(time.time())}.cali"
+            ccores = int(ncores / threads)
+            timer_cmds = f"--caliperFilename {cali_name} --adiakData 'test_name: {test_name}'"
+            finps = f"{inps} {timer_cmds}"
+            t = test(script=test_file, clas=finps,
+                     label=test_name,
+                     np=ccores,
+                     nt=threads,
+                     caliper_filename=cali_name,
+                     **kwargs)
+
+# If running CI, run gather_files on exit
+if (benchmark_dir):
+    onExit(gather_files)
+glue(keep=True, independent=True)
+
+#---------------------------------------------------------------------------
+# Test configurations
+#---------------------------------------------------------------------------
+all_tests = pt.get_all_tests()
+
+for t in all_tests:
+    spheral_setup_test(t, num_cores, num_threads)
+
+# Add a wait to ensure all timer files are done
+wait()
